Finish GET /student endpoint

+++ src/main/java/backend/model/VO/student/Score.java
@@ -12,7 +12,7 @@ import java.util.List;
 @NoArgsConstructor
 @Builder
 public class Score {
-    private Integer gradeTotal;
+    private Float gradeTotal;
 
     private List<GradeList> gradeList;
 }
